# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"aeroflot-snippets","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/api/assign":{"post":{"tags":[],"description":"Find nearest suitable engineer and assign to an aircraft","operationId":"assign_engineer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignEngineerRequest"}}},"required":true},"responses":{"200":{"description":"Successful path finding, returning a point sequence as route","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignEngineerResponse"}}}},"400":{"description":"Some error happened during task processing","content":{"text/plain":{"schema":{"type":"string"}}}},"404":{"description":"Suitable engineer was not found","content":{"text/plain":{"schema":{"type":"string"}}}},"422":{"description":"Provided JSON request has an error and cannot be parsed","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Server-side error","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/api/auth/login":{"post":{"tags":[],"description":"Login into existing user profile and get refresh and access tokens","operationId":"login_handler","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"Login failed because of credentials","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Server-side error happened","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/api/auth/register":{"post":{"tags":[],"description":"Register new user profile","operationId":"register_handler","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful registration","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"Register failed because of reques","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Server-side error happened","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/api/getRoute":{"post":{"tags":[],"description":"Classic Point-to-Point BFS","operationId":"get_route","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetRouteRequest"}}},"required":true},"responses":{"200":{"description":"Successful path finding, returning a point sequence as route","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetRouteResponse"}}}},"400":{"description":"Error caused by incorrect input data","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/api/map":{"get":{"tags":[],"operationId":"get_map","responses":{"200":{"description":"Return actual map","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MapMatrix"}}}}}}}},"components":{"schemas":{"AircraftIssue":{"type":"string","enum":["fuel_leak_from_drain_cap","oil_stain_near_gearbox","hydraulic_leak_on_strut","fairing_chip_or_scratch","paint_peeling_at_rivets","missing_pitot_cover","uneven_tread_wear","tire_cut_to_cord","low_tire_pressure","indication_fault","loose_connector","seatbelt_adjustment","burned_out_signal_lamp","thrust_or_parameter_drop","excessive_vibration","metal_debris_in_oil_filter","radar_failure_or_false_reading","comms_loss_or_distortion","ins_gyro_drift","other"]},"AssignEngineerRequest":{"type":"object","required":["dispatcher_uuid","issue","plane_point","description"],"properties":{"description":{"type":"string"},"dispatcher_uuid":{"type":"string"},"issue":{"$ref":"#/components/schemas/AircraftIssue"},"plane_point":{"$ref":"#/components/schemas/Point"}}},"AssignEngineerResponse":{"type":"object","required":["engineer_uuid","time","time_limit_exceeded","route"],"properties":{"engineer_uuid":{"type":"string"},"route":{"$ref":"#/components/schemas/Route"},"time":{"type":"number","format":"float"},"time_limit_exceeded":{"type":"boolean"}}},"EngineerType":{"type":"string","enum":["integrity_inspector","crew_remarks_handler","fueling_crew","engine_technician","avionics_engineer","aviation_technician","not_cagegorized"]},"GetRouteRequest":{"type":"object","required":["start_point","end_point"],"properties":{"end_point":{"$ref":"#/components/schemas/Point"},"start_point":{"$ref":"#/components/schemas/Point"}}},"GetRouteResponse":{"type":"object","required":["route","distance"],"properties":{"distance":{"type":"integer","minimum":0},"route":{"$ref":"#/components/schemas/Route"}}},"LoginRequest":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","required":["name","user_role","access_token","refresh_token"],"properties":{"access_token":{"type":"string"},"name":{"type":"string"},"refresh_token":{"type":"string"},"user_role":{"$ref":"#/components/schemas/UserRole"}}},"MapMatrix":{"type":"array","items":{"type":"array","items":{"type":"integer","format":"int64"}}},"Point":{"type":"array","items":{"type":"integer","format":"int64"},"maxItems":2,"minItems":2},"RegisterRequest":{"type":"object","required":["email","name","password","user_role"],"properties":{"email":{"type":"string"},"engineer_type":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/EngineerType"}]},"name":{"type":"string"},"password":{"type":"string"},"user_role":{"$ref":"#/components/schemas/UserRole"}}},"Route":{"type":"array","items":{"$ref":"#/components/schemas/Point"}},"UserRole":{"type":"string","enum":["Dispatcher","Engineer"]}}},"tags":[{"name":"Aeroflot","description":"Aeroflot dispatcher management API"}]}